Update global-datasets.rst
refining this feature.

Obs. The last picture depicting Sesam with global-datasets is not ideal. It conveys the point the writer wanted to convey, but it's highly biased since the number of inbound pipes just have decreased a lot to show the point. There should be no difference in number of inbound pipes in either of the two pictures.